Trash out

+++ trunk/Framework/AsapXna/asap/core/TickListener.cs
@@ -2,14 +2,10 @@ using System;
 
 using System.Collections.Generic;
 
-
-
 namespace asap.core
 {
     public interface TickListener
     {
-        void Tick(long delta);
-    }
-    
-    
+        void Tick(float delta);
+    }    
 }
